Revert 'scope' to 'edge'
Done in a non-git-revert way because it was easier than untangling the new tests.  Reverting because it's more intuitive, and offset() doesn't need it until I figure out how boundaries will work with offset() -- currently it'll take a moving window to evaluate a boundary formula, and I don't know how to do that nicely.