Add `drop = TRUE` argument to join_header()
This controls whether or not to drop cells that don't have a matching header.
By default, they are dropped (for backwards compatibility).

The upcoming behead() function will use `drop = FALSE` so that all cells are
retained at each stage, in case they will be subsequently be interpreted as
headers.  In the example below, if the column headers are 'beheaded' first, then
`x` and `y` must  be retained so that they can be 'beheaded' as row headers in
the next step, even though they wouldn't be attached to either of the column
headers.

```text
|   | a | b |
| x | 1 | 2 |
| y | 3 | 4 |
```