Implement `behead()` to strip one level of headers
behead() takes one level of headers from a pivot table and makes it part of the
data.  Think of it like [tidyr::gather()], except that it works when there is
more than one row of headers (or more than one column of row-headers)

This is kind of in response to #6 though it is slightly more specialised, and
hopefully more succinct, useful and comprehendable.