partition() returns corner position and other data
This is a feature change of partition.

Rather than give partitions meaningless
IDs, it returns the corner positions of the corner cells that mark each
partition.  This means that it's easy to join the corner cells.

You probably don't need to join the corner cells any more, because any columns
other than `row` and `col` will now be returned anyway, so that corners with
values can be used as headers.